[FIX] microsoft_calendar: recurrent event deletion from list view
Before this commit, deleting recurrent events from Odoo in list view mode was allowed with the synchronization active, which could trigger to spamming emails on Microsoft side. Additionaly, deleting events with the synchronization off was also deleting events on Microsoft, which should not happen.

After this commit, the deletion of recurrent events is not allowed anymore when the synchronization is active by triggering an UserError message. Also, when the synchronization is off, deleting events in Odoo don't also delete events on Microsoft side.

closes odoo/odoo#144004

Task-id: 3563501

+++ addons/microsoft_calendar/models/microsoft_sync.py
@@ -433,7 +433,7 @@ class MicrosoftSync(models.AbstractModel):
         microsoft_service = self._get_microsoft_service()
         sender_user = self._get_event_user_m(user_id)
         with microsoft_calendar_token(sender_user.sudo()) as token:
-            if token:
+            if token and not sender_user.microsoft_synchronization_stopped:
                 microsoft_service.delete(event_id, token=token, timeout=timeout)
 
     @after_commit
